+++ includes/integrations/integration-genesis-blox.php
@@ -107,3 +107,38 @@ function ddw_btc_add_tax_column_blox( $columns ) {
  * @since 1.4.0
  */
 ddw_btc_prepare_tax_column_add( 'blox' );
+
+
+if ( ! function_exists( 'ddw_genesis_tweak_plugins_submenu' ) ) :
+
+	add_action( 'admin_menu', 'ddw_genesis_tweak_plugins_submenu', 11 );
+	/**
+	 * Add Genesis submenu redirecting to "genesis" plugin search within the
+	 *   WordPress.org Plugin Directory. For Genesis 2.10.0 or higher this
+	 *   replaces the "Genesis Plugins" submenu which only lists plugins from
+	 *   StudioPress - but there are many more from the community.
+	 *
+	 * @since 1.5.1
+	 *
+	 * @uses remove_submenu_page()
+	 * @uses add_submenu_page()
+	 */
+	function ddw_genesis_tweak_plugins_submenu() {
+
+		/** Remove the StudioPress plugins submenu */
+		if ( class_exists( 'Genesis_Admin_Plugins' ) ) {
+			remove_submenu_page( 'genesis', 'genesis-plugins' );
+		}
+
+		/** Add a Genesis community plugins submenu */
+		add_submenu_page(
+			'genesis',
+			esc_html__( 'Genesis Plugins from the Plugin Directory', 'builder-template-categories' ),
+			esc_html__( 'Genesis Plugins', 'builder-template-categories' ),
+			'install_plugins',
+			esc_url( network_admin_url( 'plugin-install.php?s=genesis&tab=search&type=term' ) )
+		);
+
+	}  // end function
+
+endif;